move default pipelines of cores sources into source folders (#1888)
* move default pipelines of cores sources into source folders

* move core source templates to own folder

* rename single file template folder

* update variables

* update test imports
