+++ dlt/sources/_core_source_templates/filesystem_pipeline.py
