+++ dlt/sources/_core_source_templates/rest_api_pipeline.py
