+++ dlt/sources/_core_source_templates/sql_database_pipeline.py
