+++ dlt/sources/_single_file_templates/arrow_pipeline.py
