+++ dlt/sources/_single_file_templates/dataframe_pipeline.py
