+++ dlt/sources/_single_file_templates/default_pipeline.py
