+++ dlt/sources/_single_file_templates/fruitshop_pipeline.py
