+++ dlt/sources/_single_file_templates/github_api_pipeline.py
