+++ tests/sources/test_pipeline_templates.py
@@ -15,6 +15,6 @@ import importlib
     ],
 )
 def test_debug_pipeline(template_name: str, examples: str) -> None:
-    demo_module = importlib.import_module(f"dlt.sources.pipeline_templates.{template_name}")
+    demo_module = importlib.import_module(f"dlt.sources._single_file_templates.{template_name}")
     for example_name in examples:
         getattr(demo_module, example_name)()
